Adding TaskContext and GetTaskStates
`TaskContext` is passed to the `Task.Action` function to allow it
to change it's message while it's running. Future iterations can
potentially pass information between `Task`s via the context.

The `GetTaskStates` function pulls the `Task`'s message, status,
and depth from the leaves to the root of the task-list-tree. It
Isn't yet integrated with displaying the task-list.