Moving print from leaves to root
* Also adding `TaskContext` arg to the `Task.Skip` param
* Cleaning out old print/clear/status-indicator methods

NOTE: Adding `MaxLineLength` and `formatMessage` to `List` but
there's an unexplained error that's causing it not to work so
that feature isn't operational. There's a strange behavior when
truncating a line. The line length seems to vary before and after
the task is started. I'm not sure what significance it has on the
line length.

Also, if the `MaxLineLength` is small (~under 10 or 15) the
message behavior is very strange.