Fixing the message truncation
The error was caused by the status-indicators' `Colorizer` function.
The colorizing characters were being taken into account when calculating
line length and when truncating. This commit changes the implementation
to only count the status indicator size as 1 character and to only
truncate the message, ignoring the indent and the status indicator.

Additionally, an ellipsis character ("…") will be added to the end of
any lines that have been truncated.