Adds functionality to hide tasks when not running. Per #2
Adding the `ClearOnComplete` parameter to the `List`, to determine
if the list is printed a final time after completing and calling
the `Stop` function. The default value (`ClearOnComplete = false`)
is to keep the list.

Adding the `HideTasksWhenNotRunning` parameter to `TaskGroups` to
hide subtasks when the group is not running. This is similar to
the `List.ClearOnComplete` except that it will still show it's own
`Message` value when the subtasks are cleared and it will hide
subtasks _before_ as well as after the `TaskGroup` runs. The
difference is because the `List` doesn't have a `Message` to display
and the user can already control the pre-run list display via the
time between the `Start` and `Run` commands. The two parameters
are named differently to help suggest the difference.